combine train

+++ neural/export.py
@@ -77,29 +77,31 @@ def batch_transfer(dir):
             for name in files:
                 path1 = os.path.join(root, name)
                 path2 = os.path.join(dir_2, name)
-                shutil.copy(path1, path2)
-                image_transfer(path2)
+                if not os.path.exists(path2):
+                    shutil.copy(path1, path2)
+                    image_transfer(path2)
     else:
         print("there is not dir ", dir)
 
 
 def image_transfer(im_path):
     print(im_path)
-    img = utils.evalute_face(im_path, "./dat/79999_iter.pth", False)
-    img = utils.img_edge(img)
-    img = cv2.resize(img, (64, 64), interpolation=cv2.INTER_AREA)
-    cv2.imwrite(im_path, img)
+    if im_path.find('.jpg') >= 0:
+        img = utils.evalute_face(im_path, "./dat/79999_iter.pth", False)
+        img = utils.img_edge(img)
+        img = cv2.resize(img, (64, 64), interpolation=cv2.INTER_AREA)
+        cv2.imwrite(im_path, img)
 
 
 if __name__ == '__main__':
-    pwd = os.getcwd()
-    project_path = os.path.abspath(os.path.dirname(pwd) + os.path.sep + ".")
-    model_path = os.path.join(project_path, "unity/models/")
-    log.info(model_path)
     if len(sys.argv) > 1:
         path = sys.argv[1]
         batch_transfer(path)
     else:
+        pwd = os.getcwd()
+        project_path = os.path.abspath(os.path.dirname(pwd) + os.path.sep + ".")
+        model_path = os.path.join(project_path, "unity/models/")
+        log.info(model_path)
         shapes = [3, 4]
         for i in shapes:
             for j in range(0, 10):
